[CTRLS-78] <refactor> entity 폴더 클래스 이동 리팩터링
- 엔티티 패키지 내에 클래스가 너무 많아져서 복잡한 부분이 있어서 클래스를 따로 하위 패키지를 만들어서 성격이 겹치는 클래스를 이동시켰습니다.
단순히 리팩터링만 하였으니 따로 기능적으로 달라진 부분은 없습니다.
	modified:   src/main/java/kr/ac/hanyang/engine/Core.java
	modified:   src/main/java/kr/ac/hanyang/engine/DrawManager.java
	modified:   src/main/java/kr/ac/hanyang/engine/GameState.java
	modified:   src/main/java/kr/ac/hanyang/entity/EnemyShipSet.java
	modified:   src/main/java/kr/ac/hanyang/entity/ExperiencePool.java
	renamed:    src/main/java/kr/ac/hanyang/entity/Boss.java -> src/main/java/kr/ac/hanyang/entity/boss/Boss.java
	renamed:    src/main/java/kr/ac/hanyang/entity/Ship.java -> src/main/java/kr/ac/hanyang/entity/ship/Ship.java
	renamed:    src/main/java/kr/ac/hanyang/entity/Ship1.java -> src/main/java/kr/ac/hanyang/entity/ship/Ship1.java
	renamed:    src/main/java/kr/ac/hanyang/entity/Ship2.java -> src/main/java/kr/ac/hanyang/entity/ship/Ship2.java
	renamed:    src/main/java/kr/ac/hanyang/entity/Ship3.java -> src/main/java/kr/ac/hanyang/entity/ship/Ship3.java
	renamed:    src/main/java/kr/ac/hanyang/entity/Ship4.java -> src/main/java/kr/ac/hanyang/entity/ship/Ship4.java
	modified:   src/main/java/kr/ac/hanyang/screen/BossScreen.java
	modified:   src/main/java/kr/ac/hanyang/screen/GameScreen.java
	modified:   src/main/java/kr/ac/hanyang/screen/shipSelectScreen.java
	modified:   src/test/java/kr/ac/hanyang/entity/UltTest.java
