+++ src/main/java/kr/ac/hanyang/entity/ExperiencePool.java
@@ -3,6 +3,7 @@ package kr.ac.hanyang.entity;
 
 import java.util.HashSet;
 import java.util.Set;
+import kr.ac.hanyang.entity.ship.Ship;
 
 /**
  * Implements a pool of recyclable experience points.
